requirements.txt
certifi>=2020.6.20
requests>=2.24.0
discord.py>=1.6.0
asyncpg >= 0.22.0
invoke >= 1.5.0
